refactor: remove unnecessary NgModule, bootstrap the project using standalone APIs and use fetch() instead of XHR to read the ui config file
