+++ src/app/consent/consent.module.ts
@@ -4,7 +4,7 @@ import {ConsentDetailComponent} from "./consent-detail/consent-detail.component"
 import {ConsentDialogComponent} from "./consent-dialog/consent-dialog.component";
 import {AddConsentComponent} from "./add-consent/add-consent.component";
 import {EditConsentComponent} from "./edit-consent/edit-consent.component";
-import {SharedModule} from "../shared/shared.module";
+
 import {ConsentService} from "./consent.service";
 import {ConsentRejectedDialog} from "./dialogs/consent-rejected-dialog";
 import {ConsentInactivatedDialog} from "./dialogs/consent-inactivated-dialog";
@@ -40,16 +40,13 @@ import {
 @NgModule({
   imports: [
     CommonModule,
-    SharedModule,
     MatProgressSpinnerModule,
     MatDividerModule,
     InvalidConsentPeriodDirective,
     MatProgressBarModule,
     MatTableModule,
     EditorComponent,
-    EmptyConsentTemplateValidityPeriodDirective
-  ],
-  declarations: [
+    EmptyConsentTemplateValidityPeriodDirective,
     ConsentDetailComponent,
     ConsentDialogComponent,
     ConsentHistoryDialogComponent,
