--- src/app/shared/shared.module.ts
@@ -1,79 +1,0 @@
-import {NgModule} from '@angular/core';
-import {CommonModule} from '@angular/common';
-import {FormsModule, ReactiveFormsModule} from "@angular/forms";
-import {HasPermissionDirective} from "./directives/has-permission.directive";
-import {ErrorDialogComponent} from "./components/error-dialog/error-dialog.component";
-import {MatFormFieldModule} from "@angular/material/form-field";
-import {MatDatepickerModule} from "@angular/material/datepicker";
-import {TranslateDirective, TranslateModule, TranslatePipe} from '@ngx-translate/core';
-import {MatButtonModule} from "@angular/material/button";
-import {MatSelectModule} from "@angular/material/select";
-import {MatCardModule} from "@angular/material/card";
-import {MatRadioModule} from "@angular/material/radio";
-import {MatInputModule} from "@angular/material/input";
-import {MatDialogModule} from "@angular/material/dialog";
-import {ErrorCardComponent} from "./components/error-card/error-card.component";
-import {MatIconModule} from "@angular/material/icon";
-import {MatTooltipModule} from "@angular/material/tooltip";
-import {MatChipsModule} from "@angular/material/chips";
-import {MatAutocompleteModule} from "@angular/material/autocomplete";
-import {HasAnyPermissionsDirective} from "./directives/has-any-permissions.directive";
-import {MatSnackBarModule} from "@angular/material/snack-bar";
-import {MatExpansionModule} from "@angular/material/expansion";
-import {MatSlideToggleModule} from "@angular/material/slide-toggle";
-import {DragDropModule} from "@angular/cdk/drag-drop";
-import {
-  ConfirmDeleteDialogComponent
-} from './components/confirm-delete-dialog/confirm-delete-dialog.component';
-import {MatProgressSpinnerModule} from "@angular/material/progress-spinner";
-import {MessageCardComponent} from './components/message-card/message-card.component';
-
-@NgModule({
-  imports: [
-    CommonModule,
-    FormsModule,
-    ReactiveFormsModule,
-    MatFormFieldModule,
-    MatInputModule,
-    MatDatepickerModule,
-    MatButtonModule,
-    MatSelectModule,
-    MatCardModule,
-    MatRadioModule,
-    MatDialogModule,
-    MatIconModule,
-    MatTooltipModule,
-    MatChipsModule,
-    MatAutocompleteModule,
-    MatSnackBarModule,
-    TranslatePipe,
-    TranslateDirective,
-    MatProgressSpinnerModule
-  ],
-  declarations: [HasPermissionDirective, HasAnyPermissionsDirective, ErrorDialogComponent, ErrorCardComponent, ConfirmDeleteDialogComponent, MessageCardComponent],
-  exports: [HasPermissionDirective, HasAnyPermissionsDirective, ErrorDialogComponent, ErrorCardComponent,
-    CommonModule,
-    FormsModule,
-    ReactiveFormsModule,
-    MatFormFieldModule,
-    MatInputModule,
-    MatDatepickerModule,
-    MatButtonModule,
-    MatSelectModule,
-    MatCardModule,
-    MatRadioModule,
-    MatDialogModule,
-    MatIconModule,
-    MatTooltipModule,
-    MatChipsModule,
-    MatAutocompleteModule,
-    TranslateModule,
-    MatExpansionModule,
-    MatSlideToggleModule,
-    DragDropModule,
-    MatSnackBarModule,
-    MessageCardComponent
-  ]
-})
-export class SharedModule {
-}
